refactor(buffer): move Position struct into separate module
